Create a component for the logger safe checking
